moved parts of tests into utils and mock-data
